WP.org Themes: Further optimize the Directory's load time.
This will cut down on the amount of requests for core-bundled JS resources, and
half the size of JS app. I thought about minifying `style.css` as well but
decided to hold off on that for now. Savings are not very big and it'll
probably be more confusing than helpful.

With the project switching to maintenance mode, we can now also set a saneer
version string.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@1467 74240141-8908-4e6f-9713-ba540dce6ec7