Plugin Directory: Introduce an `approved` post status.
Approving a plugin creates the SVN repo and commits the approved plugin files
to that repo. It also add the plugin author to the list of authorised
committers and sends out a congratulatory email.

A plugin will require an initial SVN commit by the post author to then be moved
to `publish`, making it visible in the directory.

See https://wordpress.slack.com/archives/meta/p1461805883000448
See #1570.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@3036 74240141-8908-4e6f-9713-ba540dce6ec7