Plugin Directory: Add an API endpoint for generating the SVN authentication file.
`plugins.svn.wordpress.org` needs a standard SVN authentication file, which we can generate from the data stored in the `plugin_2_svn_access` table.

This endpoint is not particularly RESTful, as it dumps the file content to `STDOUT`, then `exit()`'s. This endpoint is only for internal use, so it's no big deal.

Fixes #1722.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@3295 74240141-8908-4e6f-9713-ba540dce6ec7