developer.wordpress.org: Prevent 'Requests' and 'Translations' from being treated as inline class references.
Unlike other class names, 'Requests' and 'Translations' have a decent chance of being used as actual words and not just inline class references, so err on the side of caution. If a class reference is intended for these two class names, the `@see` notation should be used, e.g. `{@see Requests}`.

Fixes #2842.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@5595 74240141-8908-4e6f-9713-ba540dce6ec7