Plugin Directory: Generate md5 hashes for plugins.
This is a POC and may change or be removed in the future, it's here for testing purposes.

A api.wordpress.org endpoint may be available in the future to access it.
This is only enabled for the 'exploit-scanner' plugin at present, purely for testing, as it publishes the md5 hashes of its own files already
 Compare https://wordpress.org/plugins/exploit-scanner/ to https://downloads.wordpress.org/plugins/exploit-scanner.1.5.2.checksums.json

See #3192


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@6022 74240141-8908-4e6f-9713-ba540dce6ec7