Official WordPress Events: Remove `sprintf()` from query.
This was only used to avoid ugly string concatenation, making the query more readible. It's too easy to confuse it with `$wpdb->prepare()`, though; a future developer could accidentally pass untrusted user input to it, expecting it to be prepared, when in reality it would be passed directly to MySQL, allowing SQL injection.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@6823 74240141-8908-4e6f-9713-ba540dce6ec7