WordCamp Gutenberg Tweaks: Avoid WP doing_it_wrong notice
In r7476 a hotfix was added to address a script localization issue that
Gutenberg was causing. However, the way the hotfix was implemented was
Doing It Wrong because `wp_script_is` was being called prior to any of the
script enqueue actions. Needless to say, with this commit, we're doing it
less wrong enough to not see that notice anymore.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@7484 74240141-8908-4e6f-9713-ba540dce6ec7