WordCamp: Prevent shutdown function from catching ignored errors
In r8260 we added detection for the @ control operator to our custom error
handler. However, the errors were still getting reported in our logs. Further
investigation showed that our custom shutdown function, which is designed to
catch and log fatal errors that would halt things before the handler could run,
was picking up and reporting the silenced error messages. The hope is that
clearing the error message that we want silenced will prevent this from
happening.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8267 74240141-8908-4e6f-9713-ba540dce6ec7