Developer: Fix parser's conversion of double underscores to strong text in parameter descriptions.
The parser interprets `__` as either as an opening or closing `<strong>`, which is never intended. This was already addressed for params via `get_params()`, but more broadly to affect other aspects of the param not just description. However, `format_param_description()` is used in other similar contexts, so this fix is needed there.

See #2900.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8466 74240141-8908-4e6f-9713-ba540dce6ec7