Session Block: Pass props to fix `attributes undefined` bug.
r8506 added support to directly changing `featured_image_width` attribute and directly uses this attribute to render image width. This caused a bbug for when `attribute` is not passed in props preventing rendering of block inside editor. This patch fixes that by passing props. Another fix is required to add sanity check in FeaturedImage component.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8519 74240141-8908-4e6f-9713-ba540dce6ec7