WordCamp: Remove obsolete helper function
The `wcorg_register_meta_only_on_endpoint` function was created to handle
registering a post meta field for only a particular post type, before Core
had a way to do so with object subtypes. Now that Core supports this, we
have stopped using this function, and it can safely be removed.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8756 74240141-8908-4e6f-9713-ba540dce6ec7