Theme Directory: Sanitize version numbers in the Theme readme.txt import process.
A number of themes include extra details in the fields such as 'WordPress', and others don't follow the plugin/theme readme stanard at all, this change validates the data a little more before importing.

It also removes the importing of the 'Tags' header as there's no immediate plans for it's use and we defer to the style.css Tags header instead.

See #3718.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@8794 74240141-8908-4e6f-9713-ba540dce6ec7