Rosetta: Showcase: Rosetta showcases aren't "public" as they don't have individual showcase pages, only a rotating section on the homepage.
This was causing Canonical to pick up showcase entries on 404 pages (ie. /five/ might attempt to redirect to a showcase site called 'Five Fingers') and cause PHP Notices.
No redirect was possible, as core correctly didn't redirect cross-domain, but it resulted in notices and a redirect to the homepage instead of a better destination.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10383 74240141-8908-4e6f-9713-ba540dce6ec7