Events API: Update the API to accept the new `type` DB field to disambiguate actual places from States/Countries/others.
As part of this the table now includes non-city locations such as countries too, and so if the best match is a country it's still handled properly (returning country-wide events rather than just distance around the central point of the country).

Updates the tests to take into account the new database changes, but these are likely to become out-of-date as the data will be updated daily going forward.

See #3728.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10474 74240141-8908-4e6f-9713-ba540dce6ec7