Theme Directory: Update the theme name & parent template when they change.
This requires that the theme name still match the slug of the theme, but allows for things like case changes and accented character changes.

If the parent theme changes, ie. a Child theme is no longer a child theme, or is now a child theme of a different theme, that's handled appropriately.
(Hopefully this rarely happens, and if it does, it's to another supported & installed theme)


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10874 74240141-8908-4e6f-9713-ba540dce6ec7