Credits: Wrap shortcode output in `<p>` for consistency.
All other text content in posts is wrapped with some kind of parent element. That's relied on in some layouts, which break if it's not present. The editor doesn't add it automatically like it does with other content, and it's impractical to add it manually.

Fixes https://github.com/WordPress/wporg-news-2021/issues/87


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11344 74240141-8908-4e6f-9713-ba540dce6ec7