Slack: Don't post a "This ticket was mentioned in.." on old closed Trac tickets.
This still posts the Ticket reference to Slack, but if the ticket has been has been closed over 2 years avoids littering it with references that were probably not about it.

See the ticket for an example of where a GitHub reference posted to a ticket that had been closed for 14 years.

Fixes #3650.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11582 74240141-8908-4e6f-9713-ba540dce6ec7