Trac: Mark links in ticket descriptions/comments as rel="ugc nofollow". Take two.
[11548] was previously applying to all 'searchable' elements which includes the code browser.
The code browser uses a different String class, and being treated as a Markup string caused it to strip content that was unexpected within it.
This implements [11548] again verifying the specific string classes which we can alter, and narrowing it further by only applying to the ticket/comment classes and only when ext-link is present, to reduce potential breakage. 

Fixes #4950.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@11583 74240141-8908-4e6f-9713-ba540dce6ec7