Photo Directory, Moderation/Rejection: Email contributor of published photo if the photo has subsequently been rejected.
The email is identical to a rejection email but with a line added "Though this photo had already been approved, we have since decided to decline it and have unpublished it from the site.". As such, a required rejection reason and optional mod note to user will be included.

Props topher1kenobe, coffee2code.
Fixes #6365.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@12107 74240141-8908-4e6f-9713-ba540dce6ec7